Update lib/package.json to include types
I encountered an issue with the vue3-plotly module where TypeScript was unable to find the type declaration file.
It appears this issue is due to the "exports" field in the package.json not including "types".
In this Pull Request, I propose adding "types" to the package.json to resolve this issue.